refactor: déplacer formatCardIdentifier et ServiceCard vers un nouveau fichier utilitaire
