+++ src/views/account/Restaurant/Cards/Card.tsx
@@ -1,5 +1,5 @@
 import { View, Text, StyleSheet, Image } from "react-native";
-import { formatCardIdentifier, ServiceCard } from "../Menu";
+import { formatCardIdentifier, ServiceCard } from "@/utils/external/restaurant";
 import { useTheme } from "@react-navigation/native";
 import { defaultProfilePicture } from "@/utils/ui/default-profile-picture";
 import { PressableScale } from "react-native-pressable-scale";
